initial support for OSX on work machine
my old config I used for this machine finally ate it. this mostly works
on the new stuff w/ a fresh nix-darwin and HM install.

I need to work out some knots and make more of the configuration modular
(with proper options and stuff), since much of it is linux specific.