shuffle into modules/{package}/default.nix
This is mostly nonfunctional, but the intent will be to import the
`glamdring.nix` module, which will pull in and default-configure
things via `common.nix`, then pull in all the other modules in the tree.

The top level can contain pre-configured `.nix` files for machines if
the config gets too unweildy to just keep in the flake.